@@ -11,42 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
-    "bootstrap": "^4.1.3",
-    "core-js": "^2.5.4",
+    "@agm/core": "^1.0.0-beta.7",
+    "@angular-devkit/build-angular": "^0.802.0",
+    "@angular/animations": "^8.1.1",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/fire": "^5.1.2",
+    "@angular/forms": "^8.1.1",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.2.0",
+    "auth0-js": "^9.11.3",
+    "bootstrap": "^4.3.1",
+    "core-js": "^2.6.9",
     "ekko-lightbox": "^5.3.0",
-    "jquery": "^3.3.1",
-    "popper.js": "^1.14.5",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.26"
+    "firebase": "^5.9.2",
+    "hammerjs": "^2.0.8",
+    "jquery": "^3.4.1",
+    "popper.js": "^1.15.0",
+    "rxjs": "~6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.5",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular/cli": "^8.2.0",
+    "@angular/compiler-cli": "^8.1.1",
+    "@angular/language-service": "^8.2.0",
+    "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "protractor": "^5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5"
   }
 }
