@@ -5,13 +5,26 @@ import { HomeComponent } from './components/home/home.component';
 import { AboutComponent } from './components/about/about.component';
 import { NavbarComponent } from './components/shared/navbar/navbar.component';
 import { ProductComponent } from './components/product/product.component';
-
+import { RegisterComponent } from './components/register/register.component';
 // import { PageNotFoundComponent } from './';
+import { SearchComponent } from './components/search/search.component';
+import { AuthGuardService } from './services/auth-guard.service';
+import { AccountComponent } from './components/account/account.component';
+import { AdminProductsComponent } from './components/admin/admin-products/admin-products.component';
+import { EditProductComponent } from './components/admin/admin-products/edit-product.component';
+import { MaterialDesignComponent } from './components/material-design/material-design.component';
+
 
 const APP_ROUTES: Routes = [
-  { path: 'home', component: HomeComponent },
+  // { path: 'material-design', component: MaterialDesignComponent },
+  { path: 'admin-products', component: AdminProductsComponent },
+  { path: 'edit-product/:id', component: EditProductComponent },
+  { path: 'home', component: HomeComponent, children: [{ path: 'material-design', component: MaterialDesignComponent }] },
   { path: 'about', component: AboutComponent },
+  { path: 'register', component: RegisterComponent },
+  { path: 'account', component: AccountComponent, canActivate: [AuthGuardService] }, // El canActivate se pone para las paginas que no queremos que se entre si no estamos logueados, lo quitaremos del register
   { path: 'product/:id', component: ProductComponent },
+  { path: 'resultado/:textoBuscador', component: SearchComponent },
   { path: '**', pathMatch: 'full', redirectTo: 'home' }
   //   { path: '**', component: PageNotFoundComponent }
 
