@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService, Product } from '../../services/products.service';
 import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
+import { log } from 'util';
+import { ChatService } from 'src/app/services/chat.service';
 
 @Component({
   selector: 'app-home',
@@ -8,12 +11,11 @@ import { Router } from '@angular/router';
 })
 export class HomeComponent implements OnInit {
   products: Product[] = [];
+  arrayDeEventos: Evento[] = [];
 
+  evento: Evento;
   // Se ejecuta antes de renderizar
-  constructor(
-    private _productsService: ProductsService,
-    private router: Router
-  ) {}
+  constructor(private _productsService: ProductsService, private router: Router, public cs: ChatService) {}
 
   // Se ejecuta cuando la página está ya renderizada
   ngOnInit() {
@@ -23,4 +25,9 @@ export class HomeComponent implements OnInit {
   // verProduct(idx: number) {
   // this.router.navigate(['/product', idx]);
   //  }
+}
+
+interface Evento {
+  nombre: string;
+  arrayDeCallBacks: any[];
 }
