@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ProductsService, Product } from '../../services/products.service';
 import { ConfigService } from '../../config/config.service';
 import { Router } from '@angular/router';
+import { timeout } from 'q';
 
 @Component({
   selector: 'app-products',
@@ -10,28 +11,41 @@ import { Router } from '@angular/router';
 })
 export class ProductsComponent implements OnInit {
   products: Product[] = [];
+  clientes: any[] = [];
   endpoint: String;
+  loading: boolean;
+  error: boolean;
 
   // Se ejecuta antes de renderizar
-  constructor(
-    private _ConfigService: ConfigService,
-    private _productsService: ProductsService,
-    private router: Router
-  ) {}
+  constructor(private _ConfigService: ConfigService, private _productsService: ProductsService, private router: Router) {
+    this.loading = true;
+    this.error = false;
+  }
 
   // Se ejecuta cuando la página está ya renderizada
   ngOnInit() {
-    this._ConfigService.getConfig().subscribe((data: String) => {
-      console.log('La url es ' + data['defaultUrl']);
-      this.endpoint = data['defaultUrl'];
-    });
-    // this._productsService
-    //   .getProducts(this.endpoint)
-    //   .subscribe((data: Product[]) => {
-    //     this.products = data['Products'];
-    //   });
-    this.products = this._productsService.getProductsOld();
-    console.log(this.products);
+    // Llamada a back
+    // this._productsService.getProducts('').subscribe(
+    //   (data: any) => {
+    //     console.log('Data es: ' + data);
+    //     this.products = data;
+    //   }, // Para manejar errores
+    //   errorServicio => {
+    //     this.error = true;
+    //     console.log(errorServicio.error.mensaje);
+    //   }
+    // );
+
+    setTimeout(() => {
+      this.products = this._productsService.getProductsOld();
+      this._productsService.getClients('').subscribe((result: any[]) => {
+        this.clientes = result;
+        console.log(this.clientes);
+      });
+      this.loading = false;
+    }, 2000);
+
+    // console.log(this.products);
   }
   verProduct(idx: number) {
     this.router.navigate(['/product', idx]);
